+++ packages/nextjs/src/index.ts
@@ -1,2 +1,2 @@
 export * from './components';
-export * from './provider';
+export * from './providers/supaglueProvider';
