+++ packages/nextjs/src/providers/index.ts
@@ -1,0 +1,2 @@
+export * from './supaglueInternalProvider';
+export * from './supaglueProvider';
