+++ packages/nextjs/src/providers/supaglueProvider.tsx
@@ -1,5 +1,3 @@
-import createCache from '@emotion/cache';
-import { CacheProvider } from '@emotion/react';
 import { createContext, FC, ReactNode, useContext, useMemo } from 'react';
 
 const SupaglueContext = createContext({
@@ -24,19 +22,6 @@ export const SupaglueProvider: FC<SupaglueProviderProps> = ({ children, customer
       {children}
     </SupaglueContext.Provider>
   );
-};
-
-type SgCacheProviderProps = {
-  children: ReactNode;
-};
-
-export const SgCacheProvider: FC<SgCacheProviderProps> = ({ children }) => {
-  const cache = createCache({
-    key: 'sg-internal',
-    // Prepending allows developers to override css since Emotion styles are at the beginning
-    prepend: true,
-  });
-  return <CacheProvider value={cache}>{children}</CacheProvider>;
 };
 
 export const useSupaglueContext = () => useContext(SupaglueContext);
